requirements.txt
flask
requests
python-dotenv
Flask-Caching
gunicorn
flasgger
